+++ common/src/main/java/net/mehvahdjukaar/amendments/integration/EtchedCompat.java
@@ -1,21 +1,50 @@
 package net.mehvahdjukaar.amendments.integration;
 
-import net.minecraft.nbt.CompoundTag;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import gg.moonflower.etched.common.component.DiscAppearanceComponent;
+import gg.moonflower.etched.core.registry.EtchedComponents;
+import net.mehvahdjukaar.amendments.AmendmentsClient;
+import net.mehvahdjukaar.moonlight.api.client.util.VertexUtil;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.util.FastColor;
 import net.minecraft.world.item.ItemStack;
 
-//TODO finish when etched updates
 public class EtchedCompat {
 
+    public static void drawDisc(ItemStack item, PoseStack poseStack, MultiBufferSource bufferSource,
+                                int lu, int lv) {
+        DiscAppearanceComponent app = (DiscAppearanceComponent)
+                item.getOrDefault(CompatObjects.DISC_APPEARANCE.get(), DiscAppearanceComponent.DEFAULT);
+        int color = app.discColor();
+        VertexConsumer builder;
+        if (color != -11447983) {
+            builder = AmendmentsClient.TINTED_RECORD.buffer(bufferSource, RenderType::entityCutout);
+            drawColoredQuad(poseStack, builder, lu, lv, color);
+        } else {
+            builder = AmendmentsClient.DEFAULT_RECORD.buffer(bufferSource, RenderType::entityCutout);
+            VertexUtil.addQuad(builder, poseStack, -0.5f, -0.5f, 0.5f, 0.5f, lu, lv);
+        }
 
-    public static int getDiscColor(ItemStack item) {
-        return 0;
+        var pattern = app.pattern();
+        if (pattern.isColorable()) {
+            builder = AmendmentsClient.RECORD_PATTERNS.get(pattern.ordinal()).buffer(bufferSource, RenderType::entityCutout);
+            int primaryColor = app.labelPrimaryColor();
+            if (primaryColor == 0) primaryColor = -1;
+            drawColoredQuad(poseStack, builder, lu, lv, primaryColor);
+            builder = AmendmentsClient.RECORD_PATTERNS_OVERLAY.get(pattern.ordinal()).buffer(bufferSource, RenderType::entityCutout);
+            int secondaryColor = app.labelSecondaryColor();
+            if (secondaryColor == 0) secondaryColor = -1;
+            drawColoredQuad(poseStack, builder, lu, lv, secondaryColor);
+        }
     }
 
-    public static CompoundTag getLabelColor(ItemStack item) {
-        return null;
-    }
-
-    public static int getPattern(ItemStack item) {
-        return 0;
+    private static void drawColoredQuad(PoseStack poseStack, VertexConsumer builder, int lu, int lv, int color) {
+        int r = FastColor.ARGB32.red(color);
+        int g = FastColor.ARGB32.green(color);
+        int b = FastColor.ARGB32.blue(color);
+        VertexUtil.addQuad(builder, poseStack, -0.5f, -0.5f, 0.5f, 0.5f,
+                r, g, b, 255, lu, lv);
     }
 }